Move Linearwarmup scheduler from base to the algo
