+++ src/otx/recipe/classification/h_label_cls/mobilenet_v3_large_light.yaml
@@ -13,7 +13,7 @@ optimizer:
     weight_decay: 0.0001
 
 scheduler:
-  - class_path: otx.core.model.module.base.LinearWarmupScheduler
+  - class_path: otx.algo.schedulers.warmup_schedulers.LinearWarmupScheduler
     init_args:
       num_warmup_steps: 10
   - class_path: lightning.pytorch.cli.ReduceLROnPlateau
