+++ src/otx/recipe/classification/h_label_cls/otx_deit_tiny.yaml
@@ -12,7 +12,7 @@ optimizer:
     weight_decay: 0.05
 
 scheduler:
-  - class_path: otx.core.model.module.base.LinearWarmupScheduler
+  - class_path: otx.algo.schedulers.warmup_schedulers.LinearWarmupScheduler
     init_args:
       num_warmup_steps: 10
   - class_path: lightning.pytorch.cli.ReduceLROnPlateau
