+++ src/otx/recipe/detection/atss_r50_fpn.yaml
@@ -10,7 +10,7 @@ optimizer:
     weight_decay: 0.0
 
 scheduler:
-  - class_path: otx.core.model.module.base.LinearWarmupScheduler
+  - class_path: otx.algo.schedulers.warmup_schedulers.LinearWarmupScheduler
     init_args:
       num_warmup_steps: 3
   - class_path: lightning.pytorch.cli.ReduceLROnPlateau
