+++ src/otx/recipe/semantic_segmentation/segnext_s.yaml
@@ -14,7 +14,7 @@ optimizer:
     weight_decay: 0.01
 
 scheduler:
-  - class_path: otx.core.model.module.base.LinearWarmupScheduler
+  - class_path: otx.algo.schedulers.warmup_schedulers.LinearWarmupScheduler
     init_args:
       num_warmup_steps: 20
   - class_path: torch.optim.lr_scheduler.PolynomialLR
